+++ packages/terra-form-select/tests/nightwatch/select/HiddenPlaceholderSelect.jsx
@@ -4,19 +4,20 @@ import Select from '../../../lib/Select';
 
 const locale = document.getElementsByTagName('html')[0].getAttribute('lang');
 
-const InvalidSelect = () => (
+const HiddenPlaceholderSelect = () => (
   <Base locale={locale}>
-    <p>This select has been marked as invalid.</p>
+    <p>Basic select with no initial selected value. The placeholder select text should be displayed.</p>
     <form>
       <Select
         name="zibby"
-        isInvalid
-        id="invalidSelect"
+        isPlaceholderHidden
+        defaultValue="puppies"
+        id="hiddenPlaceholderSelect"
       >
         <Select.Option value="puppies" display="Puppies" key="puppies" />
-        <Select.Option value="kittens" display="Kittens" key="kittens" isSelected />
+        <Select.Option value="kittens" display="Kittens" key="kittens" id="kittens" />
         <Select.Option value="snappers" display="Snappers" key="snappers" />
-        <Select.Option value="bumblers" display="Bumblers" disabled key="bumblers" />
+        <Select.Option value="bumblers" display="Bumblers" disabled key="bumblers" id="disabledOption" />
         <Select.Option value="joeys" display="Joeys" key="joeys" />
         <Select.Option value="micros" display="Microprocessors" disabled key="micros" />
       </Select>
@@ -25,4 +26,4 @@ const InvalidSelect = () => (
   </Base>
 );
 
-export default InvalidSelect;
+export default HiddenPlaceholderSelect;