+++ packages/terra-site/src/examples/form-select/SelectModalContainer.jsx
@@ -53,7 +53,7 @@ class ModalContainer extends React.Component {
             <Select.Option value="snappers" display="Snappers" key="snappers" />
             <Select.Option value="bumblers" display="Bumblers" disabled key="bumblers" />
             <Select.Option value="joeys" display="Joeys" key="joeys" />
-            <Select.Option value="micros" display="Microprocessors" disabled key="miros" />
+            <Select.Option value="micros" display="Microprocessors" disabled key="micros" />
           </Select>
         </form>
         <br />